Delete code item equipmet detail

+++ ShinobiWorld/Assets/Scripts/Bag/Equipment/EquipmentDetail.cs
@@ -83,16 +83,7 @@ namespace Assets.Scripts.Bag.Equipment
 
         public void Sell()
         {
-            Debug.Log(References.accountRefer.ID);
-            Debug.Log(accountEquipment.EquipmentID);
-            Debug.Log(Price.text);
             AccountEquipment_DAO.SellEquipment(References.accountRefer.ID, accountEquipment.EquipmentID, int.Parse(Price.text));
-
-            //References.accountRefer.Coin += int.Parse(Price.text);
-            //Player_AllUIManagement.Instance.SetUpCoinUI(References.accountRefer.Coin);
-
-            //var index = References.listAccountEquipment.FindIndex(obj => obj.EquipmentID == accountEquipment.EquipmentID);
-            //References.listAccountEquipment[index].Delete = true;
 
             References.AddCoin(int.Parse(Price.text));
 
@@ -113,20 +104,11 @@ namespace Assets.Scripts.Bag.Equipment
 
             var equip = References.listAccountEquipment.Find(obj => list.Any(filter => filter.ID == obj.EquipmentID) && obj.IsUse == true);
 
-            if (equip != null)
-            {
-                Debug.Log(equip.EquipmentID);
-
-                AccountEquipment_DAO.RemoveEquipment(References.accountRefer.ID, equip.EquipmentID);
-            }
-                Debug.Log(accountEquipment.EquipmentID);
+            if (equip != null) AccountEquipment_DAO.RemoveEquipment(References.accountRefer.ID, equip.EquipmentID);
 
             AccountEquipment_DAO.UseEquipment(References.accountRefer.ID, accountEquipment.EquipmentID);
 
-            //var index = References.listAccountEquipment.FindIndex(obj => obj.EquipmentID == accountEquipment.EquipmentID);
-            //References.listAccountEquipment[index].IsUse = true;
-
-            LoadUI();
+            Game_Manager.Instance.ReloadPlayerProperties();
 
             BagManager.Instance.ReloadEquipment(accountEquipment.EquipmentID);
         }
@@ -137,10 +119,7 @@ namespace Assets.Scripts.Bag.Equipment
 
             AccountEquipment_DAO.RemoveEquipment(References.accountRefer.ID, accountEquipment.EquipmentID);
 
-            //var index = References.listAccountEquipment.FindIndex(obj => obj.EquipmentID == accountEquipment.EquipmentID);
-            //References.listAccountEquipment[index].IsUse = false;
-
-            LoadUI();
+            Game_Manager.Instance.ReloadPlayerProperties();
 
             BagManager.Instance.ReloadEquipment(accountEquipment.EquipmentID);
         }
@@ -196,6 +175,8 @@ namespace Assets.Scripts.Bag.Equipment
                 References.accountRefer.Health -= accountEquipment.Health;
                 References.accountRefer.Chakra -= accountEquipment.Chakra;
 
+                References.UpdateAccountToDB();
+
                 AccountEquipment_DAO.UpgradeEquipment(References.accountRefer.ID, accountEquipment.EquipmentID,
                                                     DamageBonus, HealthBonus, ChakraBonus);
                 AccountEquipment_DAO.UseEquipment(References.accountRefer.ID, accountEquipment.EquipmentID);
@@ -203,17 +184,7 @@ namespace Assets.Scripts.Bag.Equipment
             else AccountEquipment_DAO.UpgradeEquipment(References.accountRefer.ID, accountEquipment.EquipmentID,
                                                     DamageBonus, HealthBonus, ChakraBonus);
 
-            //References.accountRefer.Coin -= int.Parse(CostUpgrade.text);
-            //Player_AllUIManagement.Instance.SetUpCoinUI(References.accountRefer.Coin);
-
-            var index = References.listAccountEquipment.FindIndex(obj => obj.EquipmentID == accountEquipment.EquipmentID);
-            References.listAccountEquipment[index].Health = HealthBonus;
-            References.listAccountEquipment[index].Damage = DamageBonus;
-            References.listAccountEquipment[index].Chakra = ChakraBonus;
-
-            //BagManager.Instance.ReloadEquipment(accountEquipment.EquipmentID);
             References.AddCoin(-int.Parse(CostUpgrade.text));
-            LoadUI();
             BagManager.Instance.ReloadEquipment(accountEquipment.EquipmentID);
             UpgradePanel.SetActive(false);
         }
@@ -232,31 +203,10 @@ namespace Assets.Scripts.Bag.Equipment
             AccountEquipment_DAO.DowngradeEquipment(References.accountRefer.ID, accountEquipment.EquipmentID,
                                                     equipment.Damage, equipment.Health, equipment.Chakra);
 
-            //References.accountRefer.Coin += int.Parse(CostReturn.text);
-            //Player_AllUIManagement.Instance.SetUpCoinUI(References.accountRefer.Coin);
-
-            var index = References.listAccountEquipment.FindIndex(obj => obj.EquipmentID == accountEquipment.EquipmentID);
-            References.listAccountEquipment[index].Health = equipment.Health;
-            References.listAccountEquipment[index].Damage = equipment.Damage;
-            References.listAccountEquipment[index].Chakra = equipment.Chakra;
-
-            //BagManager.Instance.ReloadEquipment(accountEquipment.EquipmentID);
             References.AddCoin(int.Parse(CostReturn.text));
             BagManager.Instance.ReloadEquipment(accountEquipment.EquipmentID);
 
             DowngradePanel.SetActive(false);
-        }
-
-        public void LoadUI()
-        {
-            //References.accountRefer.Health += Health;
-            //References.accountRefer.Chakra += Chakra;
-            Game_Manager.Instance.ReloadPlayerProperties();
-
-            Player_AllUIManagement.Instance
-                    .LoadHealthUI(References.accountRefer.Health, References.accountRefer.CurrentHealth);
-            Player_AllUIManagement.Instance
-                    .LoadChakraUI(References.accountRefer.Chakra, References.accountRefer.CurrentChakra);
         }
     }
 }
